% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangerKnockoff.R
\name{create.forest.conditionalResiduals}
\alias{create.forest.conditionalResiduals}
\title{Knockoffs with conditional residuals}
\usage{
create.forest.conditionalResiduals(
  X,
  method = "second_order",
  residuals_method = NULL,
  residuals_function = NULL,
  sigma = NULL,
  randomize = F,
  shrink = F,
  ...
)
}
\description{
Knockoffs with conditional residuals
}
