% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X}
\alias{X}
\title{rangerKnockoff Heterogeneous Data}
\format{
## `X`
A data frame with 1024 rows and 128 columns:
\describe{
 \item{0 - 95}{Numeric continuous data}
 \item{96 - 127}{Factor data}
}
}
\usage{
X
}
\description{
A set of data of mixed numeric continuous and categorical (factor) data
suitable for making knockoffs with random forests.
}
\keyword{datasets}
